+++ Asteroids/EnemyShip.cpp
@@ -7,7 +7,7 @@
 #include <iostream>
 using namespace std;
 
-const double Pi = 3.141592;
+const double Pi = (float)3.141592;
 const double Rad = Pi / 180;
 
 EnemyShip::EnemyShip(Renderer* _Rend)
@@ -23,9 +23,9 @@ EnemyShip::EnemyShip(Renderer* _Rend)
 	Width = ConFile.GetInteger("EnemyShip", "Width", 0);
 	Height = ConFile.GetInteger("EnemyShip", "Height", 0);
 	Velocity = ConFile.GetInteger("EnemyShip", "Velocity", 0);
-	HWidth = Width / 2;
-	HHeight = Height / 2;
-	H = sqrt(pow(HHeight, 2) + pow(HWidth, 2));
+	HWidth = Width / 2.0f;
+	HHeight = Height / 2.0f;
+	H = (float) sqrt(pow(HHeight, 2) + pow(HWidth, 2));
 
 
 	setNewData(true, false);
@@ -66,26 +66,26 @@ void EnemyShip::Update(Position* PlayerCenter, double PlayerHypotenuse, bool Pau
 
 		Ray.Update(Velocity, &P3);
 
-		P1.Rotation = P1.Angle * Rad;
-		P2.Rotation = P2.Angle * Rad;
-		Center.Rotation = P3.Rotation = Center.Angle * Rad;
+		P1.Rotation = (float) (P1.Angle * Rad);
+		P2.Rotation = (float) (P2.Angle * Rad);
+		Center.Rotation = P3.Rotation = (float) (Center.Angle * Rad);
 
-		P3.x = Center.x + cos(P3.Rotation) * H;
-		P3.y = Center.y - sin(P3.Rotation) * H;
-		P2.x = Center.x + cos(P2.Rotation) * H;
-		P2.y = Center.y - sin(P2.Rotation) * H;
-		P1.x = Center.x + cos(P1.Rotation) * H;
-		P1.y = Center.y - sin(P1.Rotation) * H;
+		P3.x = Center.x + cosf(P3.Rotation) * H;
+		P3.y = Center.y - sinf(P3.Rotation) * H;
+		P2.x = Center.x + cosf(P2.Rotation) * H;
+		P2.y = Center.y - sinf(P2.Rotation) * H;
+		P1.x = Center.x + cosf(P1.Rotation) * H;
+		P1.y = Center.y - sinf(P1.Rotation) * H;
 
 		if (Center.x > Rend->getWindowWidth())
 			Center.x = 0;
 		else if (Center.x < 0)
-			Center.x = Rend->getWindowWidth();
+			Center.x = (float)Rend->getWindowWidth();
 
 		if (Center.y > Rend->getWindowHeight())
 			Center.y = 0;
 		else if (Center.y < 0)
-			Center.y = Rend->getWindowHeight();
+			Center.y = (float)Rend->getWindowHeight();
 	}
 	Rend->DrawTriangle(&P1, &P2, &P3, 255, 0, 0, 255);
 }
@@ -95,11 +95,11 @@ void EnemyShip::setNewData(bool Left, bool _Active)
 	Active = _Active;
 	Ray.setActive(true);
 
-	P3.y = Center.y = Rend->getWindowHeight() / 2;
+	P3.y = Center.y = Rend->getWindowHeight() / 2.0f;
 
 	if (Left)
 	{
-		P1.x = P2.x = -Height;
+		P1.x = P2.x =  -(float)Height;
 		P3.x = 0;
 		P2.y = P3.y + HWidth;
 		P1.y = P3.y - HWidth;
@@ -107,8 +107,8 @@ void EnemyShip::setNewData(bool Left, bool _Active)
 	}
 	else
 	{
-		P1.x = P2.x = Rend->getWindowWidth() + Height;
-		P3.x = Rend->getWindowWidth();
+		P1.x = P2.x = (float)Rend->getWindowWidth() + Height;
+		P3.x = (float)Rend->getWindowWidth();
 		P2.y = P3.y - HWidth;
 		P1.y = P3.y + HWidth;
 		Center.x = P3.x + HHeight;
